cookbook/requirements.txt
backports.shutil-get-terminal-size==1.0.0
colorama==0.3.7
coverage==4.1
decorator==4.0.9
Django==1.9.7
django-nose==1.4.3
djangorestframework==3.3.3
drf-extensions==0.2.8
httpie==0.9.3
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
nose==1.3.7
pathlib2==2.1.0
pickleshare==0.7.2
Pygments==2.1.3
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
